fix(c): add signer_iid fields to lichen_frame struct
Add signer_iid, signer_iid_len, signer_iid_present fields required
by schnorr48 frame signing API. Fixes integer conversion warning
in frame.c seqnum parsing.

+++ lichen/subsys/lichen/link/frame.c
@@ -93,7 +93,7 @@ int lichen_frame_parse(struct lichen_frame *frame,
 	}
 
 	frame->epoch = data[off++];
-	frame->seqnum = ((uint16_t)data[off] << 8) | data[off + 1];
+	frame->seqnum = (uint16_t)(((uint16_t)data[off] << 8) | data[off + 1]);
 	off += 2;
 
 	/* Destination address */
